backend/requirements.txt
openai
pydantic
python-dotenv
requests
bs4
flask